remove dead code and redundant tooling
