+++ packages/bippy/package.json
@@ -102,8 +102,8 @@
     "publint": "publint",
     "test": "vp test",
     "coverage": "vp test --coverage",
-    "prepublishOnly": "cp ../../README.md . && pnpm build",
-    "build:coverage": "BIPPY_SOURCEMAP=true vp pack"
+    "prepack": "cp ../../README.md .",
+    "prepublishOnly": "nr build"
   },
   "devDependencies": {
     "@jridgewell/sourcemap-codec": "^1.5.0",
@@ -116,19 +116,16 @@
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
     "@vitest/coverage-istanbul": "^4.1.10",
-    "esbuild": "^0.27.0",
     "happy-dom": "^15.11.7",
     "ink": "^7.1.1",
     "ink-testing-library": "^4.0.0",
     "publint": "^0.3.0",
     "react": "19.0.0",
-    "react-devtools-core": "^7.0.1",
     "react-devtools-inline": "^6.0.1",
     "react-dom": "19.0.0",
     "react-nil": "^2.0.0",
     "react-refresh": "^0.16.0",
     "remotion": "^4.0.496",
-    "terser": "^5.36.0",
     "three": "^0.185.1",
     "tsx": "^4.21.0",
     "vite-plus": "latest"
