+++ packages/e2e/tests/native/source.spec.ts
@@ -1,7 +1,6 @@
 // the detox jest environment injects its own global `expect` for element
 // assertions, so jest's must be imported explicitly for plain values
 import { expect } from "@jest/globals";
-import { by, element } from "detox";
 
 import { launchFixtureApp, readElementText } from "./helpers";
