+++ packages/expo-playground/package.json
@@ -13,13 +13,11 @@
   "dependencies": {
     "@bippy/next-playground": "workspace:*",
     "bippy": "workspace:*",
-    "clsx": "^2.1.1",
     "expo": "~54.0.33",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-native": "0.81.5",
-    "react-native-web": "^0.21.2",
-    "tailwind-merge": "^2.6.0"
+    "react-native-web": "^0.21.2"
   },
   "devDependencies": {
     "@types/react": "~19.1.0",
