+++ packages/website/app/globals.css
@@ -59,7 +59,6 @@
   --text-caption--line-height: 1.25rem;
   --spacing-page: 28.1875rem;
   --spacing-faq: 26.8125rem;
-  --spacing-hero-divider: 584px;
   --spacing-intro-gap: 5px;
   --spacing-install-button: 20px;
   --spacing-faq-gap: 0.6875rem;
@@ -69,10 +68,6 @@
   --spacing-tabs-list: 3px;
   --spacing-tabs-trigger: calc(100% - 1px);
   --spacing-link-underline-offset: 5px;
-  --spacing-2.25: 0.5625rem;
-  --spacing-3.25: 0.8125rem;
-  --spacing-3.75: 0.9375rem;
-  --spacing-5.75: 1.4375rem;
   --radius-xs: 0.4375rem;
   --radius-button-xs: min(var(--radius-md), 10px);
   --radius-button-sm: min(var(--radius-md), 12px);
@@ -163,20 +158,6 @@
   --link: color(display-p3 0.1632 0.5398 0.9268);
   --link-decoration: color(display-p3 0.669 0.821 1);
   --link-decoration-hover: color(display-p3 0.48 0.66 0.92);
-  --hero-glow: linear-gradient(
-    to right,
-    transparent 0%,
-    rgba(255, 255, 255, 0.5) 15%,
-    rgba(255, 255, 255, 0.5) 85%,
-    transparent 100%
-  );
-  --hero-divider: linear-gradient(
-    to right,
-    transparent 0%,
-    rgba(0, 0, 0, 0.06) 25%,
-    rgba(0, 0, 0, 0.06) 75%,
-    transparent 100%
-  );
   --button-shadow: #0000000f 0px 0px 0px 1px, #0000000f 0px 1px 2px -1px, #0000000a 0px 2px 4px;
   --button-shadow-hover:
     #00000014 0px 0px 0px 1px, #00000014 0px 1px 2px -1px, #0000000f 0px 2px 4px;
@@ -230,14 +211,6 @@
   --link: color(display-p3 0.48 0.66 0.92);
   --link-decoration: color(display-p3 0.24 0.36 0.52);
   --link-decoration-hover: color(display-p3 0.42 0.56 0.74);
-  --hero-glow: color(display-p3 0.075 0.075 0.075);
-  --hero-divider: linear-gradient(
-    to right,
-    transparent 0%,
-    color(display-p3 1 1 1 / 0.035) 25%,
-    color(display-p3 1 1 1 / 0.035) 75%,
-    transparent 100%
-  );
   --button-shadow: #ffffff14 0px 0px 0px 1px, #00000040 0px 1px 2px -1px, #00000033 0px 2px 4px;
   --button-shadow-hover:
     #ffffff1f 0px 0px 0px 1px, #0000004d 0px 1px 2px -1px, #00000040 0px 2px 4px;
@@ -288,31 +261,5 @@
 
   .whitespace-collapse-preserve {
     white-space-collapse: preserve;
-  }
-}
-
-@layer components {
-  .hero-band-glow {
-    background: var(--hero-glow);
-  }
-
-  .hero-divider-bg {
-    background: var(--hero-divider);
-  }
-
-  .dark .shiki,
-  .dark .shiki span {
-    color: var(--shiki-dark) !important;
-  }
-
-  .fiber-viz-code .line-emphasized {
-    display: inline-block;
-    width: 100%;
-    background: color-mix(in oklab, var(--link) 9%, transparent);
-    border-radius: 3px;
-  }
-
-  .fiber-viz-code .line-dimmed {
-    opacity: 0.45;
   }
 }
