--- packages/website/components/icons/github-icon.tsx
@@ -1,10 +1,0 @@
-export const GitHubIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" {...props}>
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M5.028 21.363C3.759 20.747 2.705 19.709 1.876 18.314C1.593 17.839 1.808 17.238 2.306 16.999C2.803 16.76 3.394 16.974 3.687 17.443C4.336 18.486 5.092 19.171 5.901 19.564C6.913 20.055 8.123 20.147 9.54 19.756L10.072 21.684C8.282 22.178 6.558 22.106 5.028 21.363Z"
-    />
-    <path d="M5.893 1.25C6.324 1.25 6.924 1.351 7.33 1.449C8.05 1.624 8.972 1.958 9.833 2.606C10.582 2.43 11.643 2.223 13 2.223C14.357 2.223 15.418 2.43 16.167 2.606C17.028 1.958 17.95 1.624 18.67 1.449C19.076 1.351 19.691 1.25 20.107 1.25H20.619L20.806 1.728C21.065 2.393 21.159 3.203 21.16 3.962C21.161 4.601 21.096 5.248 20.984 5.813C21.288 6.295 21.573 6.866 21.794 7.452C22.057 8.151 22.25 8.924 22.25 9.629C22.25 12.051 21.439 13.761 20.155 14.968C19.18 15.885 17.97 16.475 16.746 16.894C17.008 17.289 17.25 17.811 17.25 18.495V20.75C17.25 21.693 17.25 22.164 16.957 22.457C16.664 22.75 16.193 22.75 15.25 22.75H10.75C9.807 22.75 9.336 22.75 9.043 22.457C8.75 22.164 8.75 21.693 8.75 20.75V18.495C8.75 17.811 8.992 17.289 9.254 16.894C8.03 16.475 6.82 15.885 5.845 14.968C4.562 13.761 3.75 12.051 3.75 9.629C3.75 8.924 3.943 8.151 4.206 7.452C4.427 6.866 4.712 6.295 5.016 5.813C4.903 5.248 4.839 4.601 4.84 3.962C4.841 3.203 4.935 2.393 5.194 1.728L5.381 1.25H5.893Z" />
-  </svg>
-);
