+++ packages/website/components/ui/tabs.tsx
@@ -60,14 +60,4 @@ const TabsTrigger = ({ className, ...props }: TabsPrimitive.Tab.Props) => {
   );
 };
 
-const TabsContent = ({ className, ...props }: TabsPrimitive.Panel.Props) => {
-  return (
-    <TabsPrimitive.Panel
-      data-slot="tabs-content"
-      className={cn("flex-1 text-sm outline-none", className)}
-      {...props}
-    />
-  );
-};
-
-export { Tabs, TabsList, TabsTrigger, TabsContent, tabsListVariants };
+export { Tabs, TabsList, TabsTrigger };
